# Switch to a faster deep clone