--- a/package.json
+++ b/package.json
@@ -29,6 +29,7 @@
     "url": "git+https://github.com/kristerkari/css-viewport-units-transform.git"
   },
   "scripts": {
+    "benchmark": "yarn build && node benchmark.js",
     "size": "npm run build && size-limit",
     "prepublish": "npm run build",
     "prebuild": "rimraf dist",
@@ -42,6 +43,7 @@
     "babel-core": "^6.26.3",
     "babel-jest": "^23.0.1",
     "babel-preset-es2015": "^6.24.1",
+    "benchmark": "^2.1.4",
     "coveralls": "^3.0.1",
     "husky": "^0.14.3",
     "jest": "^23.1.0",