Old:
-----------------------------------------
1 call x 109,049 ops/sec ±3.33% (90 runs sampled)
-----------------------------------------
4 calls x 29,072 ops/sec ±1.06% (94 runs sampled)
-----------------------------------------
8 calls x 14,634 ops/sec ±1.19% (92 runs sampled)
-----------------------------------------

New:
-----------------------------------------
1 call x 171,437 ops/sec ±2.95% (88 runs sampled)
-----------------------------------------
4 calls x 44,722 ops/sec ±1.25% (92 runs sampled)
-----------------------------------------
8 calls x 22,646 ops/sec ±1.20% (93 runs sampled)
-----------------------------------------